Add hw fb functionality to reference rasterizer
If the USE_FRAMEBUFFER environment variable is set, the reference
rasterizer will attempt to mmap 0x38000000 from /dev/mem and then copy the
result of rasterization to the 800x480x4Bpp framebuffer at that address.

If USE_FRAMEBUFFER is not set, the results of rasterization are saved
to PPM files as before.