Update index.html
Old pascal spider demo was cool. I will need to read the code to figure out how to control the spider. Right now it just wanders off the screen. It did take me a minute to figure out that "SPIDER.PAS" needed to be all caps, whereas W and R worked fine in caps or without.